requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
jinja2==3.1.4
tensorflow==2.16.2
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.5.0
aiofiles==23.2.1
